app/requirements.txt
asgiref==3.5.2
Django==4.0.5
django-model-utils==4.2.0
djangorestframework==3.13.1
lorem==0.1.1
Pillow==9.1.1
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
freezegun==1.2.1
coverage==6.4.1
